ssh: fix for KEX bug + reformats
KEX works, more or less. The first encrypted packet's first block is read
as in RFC; logged properly (packet type is unsigned byte, in Java: int).
The key derivation code is incorrect.

+++ src/main/java/me/laszloattilatoth/jada/proxy/ssh/transportlayer/PacketHandlerRegistry.java
@@ -58,7 +58,7 @@ public class PacketHandlerRegistry {
         packetHandlers[packet.packetType()].handle(packet);
     }
 
-    public void handlePacket(byte packetType, Packet packet) throws TransportLayerException {
+    public void handlePacket(int packetType, Packet packet) throws TransportLayerException {
         packetHandlers[packetType].handle(packet);
     }
